- add basic error handling for 404 - use consistent int for id, no Object needed

+++ src/main/java/com/example/asynctypicode/domain/Post.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Post {
-    private Long userId;
-    private Long id;
+    private int userId;
+    private int id;
     private String title;
     private String body;
 }
